+++ test/generator/localization_generator_test.dart
@@ -54,7 +54,7 @@ localization_gen:
         generator.generate();
 
         // Assert
-        final outputFile = File('${tempDir.path}/lib/app_localizations.dart');
+        final outputFile = File('${tempDir.path}/lib/app_localizations.gen.dart');
         expect(outputFile.existsSync(), isTrue);
 
         final content = outputFile.readAsStringSync();
@@ -91,7 +91,7 @@ localization_gen:
         expect(outputDir.existsSync(), isTrue);
 
         final outputFile =
-            File('${tempDir.path}/lib/generated/app_localizations.dart');
+            File('${tempDir.path}/lib/generated/app_localizations.gen.dart');
         expect(outputFile.existsSync(), isTrue);
       });
 
@@ -139,7 +139,38 @@ localization_gen:
         generator.generate();
 
         final outputFile =
-            File('${tempDir.path}/lib/my_custom_localizations.dart');
+            File('${tempDir.path}/lib/my_custom_localizations.gen.dart');
+        expect(outputFile.existsSync(), isTrue);
+      });
+
+      test('supports custom output_file_suffix', () {
+        final localizationsDir = Directory('${tempDir.path}/localizations');
+        localizationsDir.createSync();
+        final outputDir = Directory('${tempDir.path}/lib');
+        outputDir.createSync();
+
+        TestHelper.createJsonFile(
+          localizationsDir,
+          'app_en.json',
+          TestHelper.basicEnglishJson(),
+        );
+
+        final configFile = File('${tempDir.path}/pubspec.yaml');
+        configFile.writeAsStringSync('''
+name: test_app
+localization_gen:
+  input_dir: ${tempDir.path}/localizations
+  output_dir: ${tempDir.path}/lib
+  class_name: AppLocalizations
+  output_file_suffix: .dart
+''');
+
+        final generator = LocalizationGenerator(
+          configPath: configFile.path,
+        );
+        generator.generate();
+
+        final outputFile = File('${tempDir.path}/lib/app_localizations.dart');
         expect(outputFile.existsSync(), isTrue);
       });
     });